requirements.txt
aiopg==1.4.0
async-timeout==4.0.2
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
peewee==3.16.0
peewee-async==0.8.0
psycopg2-binary==2.9.5
py-bcrypt==0.4
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
validator==0.7.1
Werkzeug==2.2.3
